+++ test/bin/browser_aggregate_tests_test.dart
@@ -6,8 +6,8 @@ import 'package:test_descriptor/test_descriptor.dart' as d;
 import 'package:test_process/test_process.dart';
 
 void main() {
-  Future<String> createProject({bool dart2jsAggregation}) async {
-    dart2jsAggregation ??= true;
+  Future<String> createProject({bool browserAggregation}) async {
+    browserAggregation ??= true;
     await d.dir('pkg', [
       d.file('pubspec.yaml', '''name: pkg
 environment:
@@ -20,16 +20,16 @@ dev_dependencies:
   test_html_builder:
     path: ${p.current}
 '''),
-      if (dart2jsAggregation) d.file('build.yaml', '''targets:
+      if (browserAggregation) d.file('build.yaml', '''targets:
   \$default:
     builders:
       test_html_builder:
         options:
-          dart2js_aggregation: true
+          browser_aggregation: true
 '''),
-      if (dart2jsAggregation)
+      if (browserAggregation)
         d.file(
-            'dart_test.yaml', 'include: test/dart_test.dart2js_aggregate.yaml'),
+            'dart_test.yaml', 'include: test/dart_test.browser_aggregate.yaml'),
       d.dir('test', [
         d.file('foo_test.dart', '''@TestOn('browser')
 import 'package:test/test.dart';
@@ -42,24 +42,24 @@ void main() {
     return d.path('pkg');
   }
 
-  Future<TestProcess> testDart2jsAggregateExecutable(List<String> args,
+  Future<TestProcess> testBrowserAggregateExecutable(List<String> args,
       {String workingDirectory}) async {
     final pubGet = await TestProcess.start('pub', ['get'],
         workingDirectory: workingDirectory);
     await pubGet.shouldExit(0);
     return TestProcess.start(
-        'pub', ['run', 'test_html_builder:dart2js_aggregate_tests', ...args],
+        'pub', ['run', 'test_html_builder:browser_aggregate_tests', ...args],
         workingDirectory: workingDirectory);
   }
 
   test('--mode=args', () async {
     final dir = await createProject();
-    final process = await testDart2jsAggregateExecutable(['--mode=args'],
+    final process = await testBrowserAggregateExecutable(['--mode=args'],
         workingDirectory: dir);
     expect(
         process.stdout,
         emitsInOrder([
-          '--release --build-filter=test/templates/default_template.dart2js_aggregate_test.** -- --preset=dart2js-aggregate',
+          '--build-filter=test/templates/default_template.browser_aggregate_test.** -- --preset=browser-aggregate',
           emitsDone
         ]));
     await process.shouldExit(0);
@@ -67,19 +67,19 @@ void main() {
 
   test('--mode=build', () async {
     final dir = await createProject();
-    final process = await testDart2jsAggregateExecutable(['--mode=build'],
+    final process = await testBrowserAggregateExecutable(['--mode=build'],
         workingDirectory: dir);
     expect(
         process.stdout,
         emitsInOrder([
-          emitsThrough('Building dart2js aggregate test config...'),
+          emitsThrough('Building browser aggregate test config...'),
           emitsThrough(
-              'pub run build_runner build --build-filter=test/dart_test.dart2js_aggregate.yaml'),
+              'pub run build_runner build --build-filter=test/dart_test.browser_aggregate.yaml'),
           emitsThrough(contains('Succeeded')),
-          emitsThrough('Reading dart2js aggregate test config...'),
+          emitsThrough('Reading browser aggregate test config...'),
           emitsThrough('Found 1 aggregate tests to run.'),
           emitsThrough(
-              'pub run build_runner build --release --build-filter=test/templates/default_template.dart2js_aggregate_test.**'),
+              'pub run build_runner build --build-filter=test/templates/default_template.browser_aggregate_test.**'),
           emitsThrough(contains('Succeeded')),
         ]));
     await process.shouldExit(0);
@@ -89,45 +89,45 @@ void main() {
     final dir = await createProject();
     // --mode=test is the default
     final process =
-        await testDart2jsAggregateExecutable([], workingDirectory: dir);
+        await testBrowserAggregateExecutable([], workingDirectory: dir);
     expect(
         process.stdout,
         emitsInOrder([
-          emitsThrough('Building dart2js aggregate test config...'),
+          emitsThrough('Building browser aggregate test config...'),
           emitsThrough(
-              'pub run build_runner build --build-filter=test/dart_test.dart2js_aggregate.yaml'),
+              'pub run build_runner build --build-filter=test/dart_test.browser_aggregate.yaml'),
           emitsThrough(contains('Succeeded')),
-          emitsThrough('Reading dart2js aggregate test config...'),
+          emitsThrough('Reading browser aggregate test config...'),
           emitsThrough('Found 1 aggregate tests to run.'),
           emitsThrough(
-              'pub run build_runner test --release --build-filter=test/templates/default_template.dart2js_aggregate_test.** -- --preset=dart2js-aggregate'),
+              'pub run build_runner test --build-filter=test/templates/default_template.browser_aggregate_test.** -- --preset=browser-aggregate'),
           emitsThrough(contains('All tests passed!')),
         ]));
     await process.shouldExit(0);
   });
 
-  test('debug', () async {
+  test('release', () async {
     final dir = await createProject();
-    final process = await testDart2jsAggregateExecutable(
-        ['--mode=args', '--debug'],
+    final process = await testBrowserAggregateExecutable(
+        ['--mode=args', '--release'],
         workingDirectory: dir);
     expect(
         process.stdout,
         emitsInOrder([
-          '--build-filter=test/templates/default_template.dart2js_aggregate_test.** -- --preset=dart2js-aggregate',
+          '--release --build-filter=test/templates/default_template.browser_aggregate_test.** -- --preset=browser-aggregate',
           emitsDone
         ]));
     await process.shouldExit(0);
   });
 
-  test('warns when dart2js aggregation is not enabled', () async {
-    final dir = await createProject(dart2jsAggregation: false);
-    final process = await testDart2jsAggregateExecutable(['--mode=args'],
+  test('warns when browser aggregation is not enabled', () async {
+    final dir = await createProject(browserAggregation: false);
+    final process = await testBrowserAggregateExecutable(['--mode=args'],
         workingDirectory: dir);
     expect(
         process.stdout,
         emits(contains(
-            'dart2js aggregation is not enabled. Update your build.yaml')));
+            'browser aggregation is not enabled. Update your build.yaml')));
     await process.shouldExit(isNot(0));
   });
 }